Add ability to restrict children.
Added child_types list.

Cache child types using a class variable to ensure that get_child_types is run
once per page class, per django initiation.

The list is list of values similar to those values accepted in a model
ForeignKey.

i.e. ["self", "app.Model", "Model", Model]
With "self" referring to the model on which can_have_children is defined.
"app.Model" is an app_label-model dot string.
"Model" is a model in the current app.
And the last is a Model class or Model instance.